added nav and cards component
Ï separate card and nav component in app.js
